requirements_new_virtualenv.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.1
gast==0.2.2
grpcio==1.16.1
h5py==2.8.0
idna==2.8
imageio==2.5.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras-vis==0.4.1
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
mock==2.0.0
networkx==2.4
nose==1.3.7
numpy==1.16.3
olefile==0.46
pbr==5.1.3
Pillow==6.0.0
protobuf==3.7.1
pycparser==2.19
pydot==1.4.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.1
requests==2.21.0
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==6.0.2
tqdm==4.31.1
urllib3==1.24.2
Werkzeug==0.15.2
